EDAnalyzer --> EDProducer; name change; dummy output
